feat: Add primary java objects for TicTacToe game board
The TicTacToeBoard holds a reference to a two-dimensional array
of TicTacToeTiles, which is instantiated with a given row size.
The board will always be an even square of x tiles.

Each TicTacToeTile holds a reference to its row index and column index
in the gameboard, and current state, which defines whether or not
the tile is open or occupied by a certain player.