feat: Add WinCondition models
When the TicTacToeBoard is generated, the presenter will need to generate
win conditions based on the size of the board and the game rules.

These WinCondition models will be used by the presenter to evaluate
whether or not the last player's move was a game-winning move.

The Callback allows handling for three events: if the win condition
has been met, which results in a game over, if the condition still has
an open tile, which means the condition hasn't been met yet, or if the
condition contains tiles occupied by both Player X and O, which means
the condition will never be met at that point.

WinConditionTypes captures the various win condition categories as defined
in the rules, namely a completed row, column, diagonal, 2x2 square, or
the board corners.