feat: Add utility class to help generate win conditions
This class will handle the win condition instantiations based on a
gameboard of TicTacToe tiles passed to the static generate method.

This returns a SparseArrayCompat map object, mapping each grid index
with a list of possible WinConditions associated with that tile index.
By ordering this in a map, this allows the presenter to only evaluate
win conditions that have been affected by the last player move.