feat: Add game layout files
Create versions of the activity and fragment for both portrait
and landscape orientation.  The fragment holds the gameboard
content, and could be changed out for other similar-styles
gameboards, if necessary.  The game_tile_item holds the imageview
for each tictactoe tile, which ultimately displays the current status
of the tile.