refactor: Better handle Activity game status UI updates
Add another member variable, winningPlayer, which is set to OPEN
until a player wins the game.  Then, during the launchNewTicTacToe
method, check to see if the game is over.  If not, have the View
display the current player.  If it is, see if the winningPlayer
is still set to OPEN.  If it is, the game was a draw, and if not,
the game was won by the set winning player.  In either case, have the
view display the current status.

This allows us to remove the onSavedInstanceState/onRestoreInstanceState
from the GameActivity, which was cluttered and unnecessary, and gives
the Presenter a stronger responsibility for keeping the View UI updated,
as opposed to leaving responsibilities to the View components.